Don't force block editor

+++ class-imagely-editor.php
@@ -43,7 +43,7 @@ if (!class_exists('Imagely_Editor')) {
             if (class_exists('Classic_Editor')) {
                 $klass = new ReflectionClass('Classic_Editor');
                 $version = $klass->getConstant('plugin_version');
-                if (version_compare($version, "0.5") > 0) {
+                if (version_compare($version, "0.5") > 0 && get_option('classic-editor-replace') == 'no-replace') {
                     update_option('classic-editor-replace', 'block');
                 }
             }
